failed attempt at re-partitioning dangling comments
see https://github.com/astral-sh/ruff/pull/21410#issuecomment-3534322840, but
the short summary is that `if` (and likely other) statement formatting code that
uses `maybe_parenthesize` checks if the condition has any leading or trailing
comments, so if we try to smuggle the comments in as dangling comments, it
thinks the expression won't break, so it doesn't add parentheses when formatting
a case like this:

```py
if (
  not
  # comment
  a):
    pass
```

and we end up with a syntax error:

```py
if
not a:
    pass
```

There may be some other way around this, but this is why I'm giving up for now.
It really feels like we want another CommentPlacement variant or some kind of
dangling tag, like Micha mentioned when I met with him.