Fix default test executable in py-fuzzer
Summary
--

I was firing up the fuzzer tonight and hit an assertion error here. We now build
with the `profiling` profile, so we need to use that executable too.

This hasn't affected CI because we always set the `--test-executable`.

Test Plan
--

Ran the script again with the same arguments on this branch