[ty] suppress suggestions during variable binding
Statements such as `def foo(p<CURSOR>`,
`def foo[T<CURSOR>` and `for foo<CURSOR>`
should not generate any suggestions as these
cases are introducing new names.

If it's not possible to determine that suggestions
should be omitted using token matching, we turn
to traversing the AST to determine the context.