Set the diagnostic URL for lint errors
Summary
--

This PR wires up the `Diagnostic::set_documentation_url` method from #21502 to
Ruff's lint diagnostics and renders them in the grouped output format (concise
and full are already covered by #21502).

I considered also including the URLs for the pylint output format, but it
doesn't currently render any color in the CLI, so it may be a machine-readable
format too?

Test Plan
--