[flake8-commas] Add an option for allowing single-argument functions (COM812).
# Summary

Resolves #18258

The following patterns currently are incompatible with enabling the
COM812 rule:

```python
some_string = gettext(
    'This is a long string which '
    'is marked for internationalization'
)

some_function(
    a_long +
    expression
)
```

The fundamental intent for COM812 is to minimize diffs when adding
additional entries to a list, dict, or function call. In the case of the
above, without an option, users are limited to one of these options:

```python
# Include a trailing comma. This is ugly, and if for some reason another
# sentence was added, the comma has to be removed.
s = gettext(
    'This is a long string '
    'which is marked for i18n.',
)

# Put the closing `)` on the same line. Adding another argument to this
# requires changing the `)` to a `,`, so there's no minimization of the
# diff here either.
some_function(
    a_long +
    expression)


# Assigning an expression to a variable and then calling the function
# works in some cases, but it bloats the code, and explicitly cannot
# work in the case of `gettext` where the scanner that extracts strings
# for the .po files doesn't know much about Python, and can only extract
# string literals inside of the `gettext(...)` calls.
var = (
    a_long +
    expression
)
some_function(var)

broken = (
    'This string will not be properly '
    'extracted by gettext scanners.'
)
s = gettext(broken)
```

# Resolution

This change adds a new configuration key:
```toml
[lint.flake8-commas]
allow_single_arg_function_calls = true
```

Setting this allow the COM812 rule to flag everything it normally does,
but with an exception for a function which is called with a single
argument.


# Tests

A new test case has been added which uses the existing COM81* test
fixture, and the resulting snapshot has been carefully reviewed to
verify correctness. One additional case has been added to that file to
verify that single-element dictionaries are not affected.