[ty] Fix range calculation in variable binding completions
Previously the cursor position was calculated using the length
of the typed token even if the cursor position was in the
middle of the token (f<CURSOR>oo).

I haven't managed to construct a test based on autocomplete suggestions
that would fail because of this but if you test the
`is_in_variable_binding` function directly it becomes clear
that the previous calculation was not entirely correct.